requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
httptools>=0.6.1
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
pytz>=2023.3
pydantic>=2.4.2
pydantic-settings>=2.0.0
google-auth>=2.23.4
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.108.0
slack-sdk>=3.22.0
openai>=1.3.7
ollama>=0.1.7
transformers>=4.30.0
torch>=2.0.0
SpeechRecognition>=3.8.1
pyttsx3>=2.90
sounddevice>=0.4.6
scipy>=1.11.4
numpy>=1.25.2
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
aiohttp>=3.8.0
cryptography>=41.0.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
aiofiles>=23.2.1
websockets>=12.0
requests>=2.31.0
python-dateutil>=2.8.2
email-validator>=2.1.0
google-generativeai>=0.3.0 
PyYAML>=6.0 